+++ dubbo-samples-stub/src/main/java/org/apache/dubbo/samples/stub/EmbeddedZooKeeper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.samples.stub;
+package org.apache.dubbo.samples.stub;
 
 import java.io.File;
 import java.lang.reflect.Method;
