+++ dubbo-samples-stub/src/main/java/org/apache/dubbo/samples/stub/impl/DemoServiceStub.java
@@ -17,21 +17,28 @@
  *
  */
 
-package com.alibaba.dubbo.samples.stub;
+package org.apache.dubbo.samples.stub.impl;
 
-import com.alibaba.dubbo.samples.stub.api.DemoService;
+import org.apache.dubbo.samples.stub.api.DemoService;
 
-import org.springframework.context.support.ClassPathXmlApplicationContext;
+public class DemoServiceStub implements DemoService {
 
-public class StubConsumer {
+    private final DemoService demoService;
 
-    public static void main(String[] args) {
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(
-            new String[] {"spring/stub-consumer.xml"});
-        context.start();
-        DemoService demoService = (DemoService)context.getBean("demoService"); // get remote service proxy
-        demoService.sayHello("aaa");
+    public DemoServiceStub(DemoService demoService) {
+        this.demoService = demoService;
+    }
 
+    @Override
+    public String sayHello(String name) {
+        //client check code goes here
+        System.out.println("stub sayHello");
 
+        try {
+            return demoService.sayHello(name);
+        } catch (Exception e) {
+           //handle Exception
+            return null;
+        }
     }
 }
