+++ dubbo-samples-stub/src/main/resources/spring/stub-consumer.xml
@@ -33,6 +33,6 @@
 
     <!-- generate proxy for the remote service, then demoService can be used in the same way as the
     local regular interface -->
-    <dubbo:reference id="demoService" check="false" interface="com.alibaba.dubbo.samples.stub.api.DemoService"/>
+    <dubbo:reference id="demoService" check="false" interface="org.apache.dubbo.samples.stub.api.DemoService"/>
 
 </beans>
