+++ dubbo-samples-stub/src/main/resources/spring/stub-provider.xml
@@ -34,9 +34,9 @@
     <dubbo:protocol name="dubbo" port="20880"/>
 
     <!-- service implementation, as same as regular local bean -->
-    <bean id="demoService" class="com.alibaba.dubbo.samples.stub.impl.DemoServiceImpl"/>
+    <bean id="demoService" class="org.apache.dubbo.samples.stub.impl.DemoServiceImpl"/>
 
     <!-- declare the service interface to be exported -->
-    <dubbo:service interface="com.alibaba.dubbo.samples.stub.api.DemoService" ref="demoService" stub="com.alibaba.dubbo.samples.stub.impl.DemoServiceStub"/>
+    <dubbo:service interface="org.apache.dubbo.samples.stub.api.DemoService" ref="demoService" stub="org.apache.dubbo.samples.stub.impl.DemoServiceStub"/>
 
 </beans>
